move ghosted_mobile_units and highlight_geos from widget to gadget

+++ luarules/gadgets/gui_ghosted_mobile_units.lua
@@ -1,11 +1,7 @@
-include("colors.h.lua")
-include("keysym.h.lua")
-local versionNumber = "1.21"
-
-function widget:GetInfo()
+function gadget:GetInfo()
     return {
         name      = "Ghost Mobile Units",
-        desc      = "Displays ghosted mobile units under their radar blips",
+        desc      = "Displays ghosted mobile enemy units under their radar blips\n(the engine does it automatically for immobile units)",
         author    = "very_bad_soldier",
         date      = "July 21, 2008",
         license   = "GNU GPL v2",
@@ -14,24 +10,31 @@ function widget:GetInfo()
     }
 end
 
+if  (gadgetHandler:IsSyncedCode()) then
+    return false
+end
+
 local lineWidth = 1.0 -- calcs dynamic now
 
 local printDebug
 
-local udefTab                = UnitDefs
-local spGetUnitDefID        = Spring.GetUnitDefID
-local spEcho                = Spring.Echo
-local spGetUnitPosition     = Spring.GetUnitPosition
-local spGetGameSeconds      = Spring.GetGameSeconds
-local floor                 = math.floor
-local pairs                    = pairs
-local spGetMyPlayerID       = Spring.GetMyPlayerID
-local spGetPlayerInfo       = Spring.GetPlayerInfo
-local spIsUnitInView        = Spring.IsUnitInView
+local udefTab                 = UnitDefs
+local spGetUnitDefID          = Spring.GetUnitDefID
+local spEcho                  = Spring.Echo
+local spGetUnitPosition       = Spring.GetUnitPosition
+local spGetGameSeconds        = Spring.GetGameSeconds
+local floor                   = math.floor
+local pairs                   = pairs
+local spGetMyPlayerID         = Spring.GetMyPlayerID
+local spGetPlayerInfo         = Spring.GetPlayerInfo
+local spIsUnitInView          = Spring.IsUnitInView
+local spIsUnitAllied          = Spring.IsUnitAllied
+local spGetUnitPosErrorParams = Spring.GetUnitPosErrorParams
+local spGetRadarErrorParams   = Spring.GetRadarErrorParams
 
 local glColor               = gl.Color
 local glDepthTest           = gl.DepthTest
-local glUnitShape            = gl.UnitShape
+local glUnitShape           = gl.UnitShape
 local glPopMatrix           = gl.PopMatrix
 local glPushMatrix          = gl.PushMatrix
 local glTranslate           = gl.Translate
@@ -42,20 +45,36 @@ local dots = {}
 local lastTime
 local updateInt = 2 --seconds for the ::update loop
 
-function widget:UnitEnteredRadar(unitID, allyTeam)
+local myTeamID = Spring.GetMyTeamID()
+local allyTeamID = Spring.GetMyAllyTeamID()
+
+function gadget:PlayerChanged()
+    myTeamID = Spring.GetMyTeamID()
+    allyTeamID = Spring.GetMyAllyTeamID()
+end
+
+function gadget:UnitEnteredRadar(unitID, allyTeam)
+    if spIsUnitAllied(unitID) then
+        return
+    end
+
     if ( dots[unitID] ~= nil ) then
         dots[unitID]["radar"] = true
     end
 end
 
-function widget:UnitEnteredLos(unitID, allyTeam )
+function gadget:UnitEnteredLos(unitID, allyTeam )
+    if spIsUnitAllied(unitID) then
+        return
+    end
+
     --update unitID info, ID could have been reused already!
     local udefID = spGetUnitDefID(unitID)
     local udef = udefTab[udefID]
         
     -- skip buildings, they get drawn ghosted anyway by the engine
     -- but mobile units don't
-    -- (but both buildings and mobile units do get 'ghosted' radar dots)
+    -- (both buildings and mobile units do get 'ghosted' radar dots)
     if ( udef.isBuilding == false and udef.isFactory == false ) then 
         dots[unitID] = {}
         dots[unitID]["unitDefId"] = udefID
@@ -68,43 +87,60 @@ function widget:UnitEnteredLos(unitID, allyTeam )
 end
 
 
-function widget:UnitCreated(unitID, allyTeam)
+function gadget:UnitCreated(unitID, allyTeam)
+    if spIsUnitAllied(unitID) then
+        return
+    end
+
     --kill the dot info if this unitID gets reused on own team
     if ( dots[unitID] ~= nil ) then
         dots[unitID] = nil
     end
 end
 
-function widget:UnitDestroyed(unitID, unitDefID, teamID, attackerID, attackerDefID, attackerTeamID)
+function gadget:UnitDestroyed(unitID, unitDefID, teamID, attackerID, attackerDefID, attackerTeamID)
+    if spIsUnitAllied(unitID) then
+        return
+    end
+
     if ( dots[unitID] ~= nil ) then
         dots[unitID] = nil
     end
 end
 
-function widget:UnitLeftRadar(unitID, allyTeam)
+function gadget:UnitLeftRadar(unitID, allyTeam)
+    if spIsUnitAllied(unitID) then
+        return
+    end
+
     if ( dots[unitID] ~= nil ) then
         dots[unitID]["radar"] = false
     end
 end
 
-function widget:UnitLeftLos(unitID, allyTeam)
+function gadget:UnitLeftLos(unitID, allyTeam)
+    if spIsUnitAllied(unitID) then
+        return
+    end
+
     if ( dots[unitID] ~= nil ) then
         dots[unitID]["los"] = false
     end
 end
 
 
-function widget:DrawWorld()
+function gadget:DrawWorld()
     glColor(1.0, 1.0, 1.0, 0.35 )
     glDepthTest(true)
 
     for unitID, dot in pairs( dots ) do
-        Spring.Echo(0, dot["radar"] , dot["los"] , dot["unitDefId"])
         if ( dot["radar"] == true ) and ( dot["los"] == false ) and ( dot["unitDefId"] ~= nil ) then
             local x, y, z = spGetUnitPosition(unitID)
+            local ex, ey, ez = spGetUnitPosErrorParams(unitID)
+            local allyteamErrorSize,_,_ = spGetRadarErrorParams(allyTeamID)
             if x and ( spIsUnitInView(unitID) ) then
                 glPushMatrix()
-                glTranslate( x, y + 5 , z)
+                glTranslate( x+ex*allyteamErrorSize, y+ey*allyteamErrorSize + 5 , z+ez*allyteamErrorSize )
                 glUnitShape( dot["unitDefId"], dot["teamId"], false, true, false)                          
                 glPopMatrix()
             end
@@ -115,7 +151,7 @@ function widget:DrawWorld()
      glColor(1, 1, 1, 1)
 end
 
-function widget:Update()
+function gadget:Update()
     local timef = spGetGameSeconds()
     local time = floor(timef)
 
@@ -127,7 +163,7 @@ function widget:Update()
         local _, _, spec, _, _, _, _, _ = spGetPlayerInfo(playerID)
         
         if ( spec == true ) then
-            widgetHandler:RemoveWidget()
+            gadgetHandler:RemoveGadget()
             return false
         end
     end
